Support https connections.
Instead of setuping a dual server for http and https, we will use the
heroku reverse proxy for https connections. This mean we no longer
support http. The old parameters hostname, path and port are now
combined in the new url parameter. From the protocol field in the URI we
know what kind of request we must do to the remote site.

We also enforce ssl connections on production environment through
express-sslify.